--- src/bin/recognize.py
@@ -1,104 +1,0 @@
-#!/usr/bin/env python
-import argparse
-import json
-import time
-import torch
-import kaldi_io
-
-from transformer import Transformer
-from utils import add_results_to_json, process_dict
-from data import build_LFR_features
-
-parser = argparse.ArgumentParser(
-    "End-to-End Automatic Speech Recognition Decoding.")
-# data
-parser.add_argument('--recog-json', type=str, required=True,
-                    help='Filename of recognition data (json)')
-parser.add_argument('--dict', type=str, required=True,
-                    help='Dictionary which should include <unk> <sos> <eos>')
-parser.add_argument('--result-label', type=str, required=True,
-                    help='Filename of result label data (json)')
-# model
-parser.add_argument('--model-path', type=str, required=True,
-                    help='Path to model file created by training')
-# decode
-parser.add_argument('--beam-size', default=1, type=int,
-                    help='Beam size')
-parser.add_argument('--nbest', default=1, type=int,
-                    help='Nbest size')
-parser.add_argument('--print-freq', default=1, type=int,
-                    help='print_freq')
-parser.add_argument('--decode-max-len', default=0, type=int,
-                    help='Max output length. If ==0 (default), it uses a '
-                    'end-detect function to automatically find maximum '
-                    'hypothesis lengths')
-
-
-def recognize(args):
-    model, LFR_m, LFR_n = Transformer.load_model(args.model_path)
-    print(model)
-    model.eval()
-    model.cuda()
-    char_list, sos_id, eos_id = process_dict(args.dict)
-    assert model.decoder.sos_id == sos_id and model.decoder.eos_id == eos_id
-
-    # read json data
-    with open(args.recog_json, 'rb') as f:
-        js = json.load(f)['utts']
-
-    # decode each utterance
-    new_js = {}
-    with torch.no_grad():
-        for idx, name in enumerate(js.keys(), 1):
-            print('(%d/%d) decoding %s' %
-                  (idx, len(js.keys()), name), flush=True)
-            input = kaldi_io.read_mat(js[name]['input'][0]['feat'])  # TxD
-            input = build_LFR_features(input, LFR_m, LFR_n)
-            input = torch.from_numpy(input).float()
-            input_length = torch.tensor([input.size(0)], dtype=torch.int)
-            input = input.cuda()
-            input_length = input_length.cuda()
-            nbest_hyps = model.recognize(input, input_length, char_list, args)
-            new_js[name] = add_results_to_json(js[name], nbest_hyps, char_list)
-
-    with open(args.result_label, 'wb') as f:
-        f.write(json.dumps({'utts': new_js}, indent=4,
-                           sort_keys=True).encode('utf_8'))
-
-
-def fast_recognize(args):
-    from data import AudioDataLoader, AudioDataset
-
-    cv_dataset = AudioDataset(args.recog_json, args.batch_size,
-                              args.maxlen_in, args.maxlen_out,
-                              batch_frames=args.batch_frames)
-
-    cv_loader = AudioDataLoader(cv_dataset, batch_size=1,
-                                num_workers=args.num_workers,
-                                LFR_m=args.LFR_m, LFR_n=args.LFR_n)
-
-    model, LFR_m, LFR_n = Transformer.load_model(args.model_path)
-    print(model)
-    model.eval()
-    model.cuda()
-    char_list, sos_id, eos_id = process_dict(args.dict)
-    assert model.decoder.sos_id == sos_id and model.decoder.eos_id == eos_id
-
-    for i, data in enumerate(cv_loader):
-        start = time.time()
-        padded_input, input_lengths, padded_target = data
-        padded_input = padded_input.cuda()
-        input_lengths = input_lengths.cuda()
-        padded_target = padded_target.cuda()
-        pred, gold = model.recognize(padded_input, input_lengths, padded_target)
-
-        if i % args.print_freq == 0:
-            print('Iter {} | {:.1f} ms/batch'.
-                  format(i + 1, 1000 * (time.time() - start) / (i + 1)),
-                  flush=True)
-
-
-if __name__ == "__main__":
-    args = parser.parse_args()
-    print(args, flush=True)
-    recognize(args)
