+++ src/ctcModel/solver.py
@@ -1,143 +1,13 @@
-import os
 import time
-
 import torch
 
-from ctcModel.loss import cal_performance
-from utils.utils import IGNORE_ID
+from ctcModel.loss import cal_loss
+from utils.solver import Solver
 
 
-class Solver(object):
+class CTC_Solver(Solver):
     """
     """
-
-    def __init__(self, data, model, optimizer, args):
-        self.tr_loader = data['tr_loader']
-        self.cv_loader = data['cv_loader']
-        self.model = model
-        self.optimizer = optimizer
-
-        # Low frame rate feature
-        self.LFR_m = args.LFR_m
-        self.LFR_n = args.LFR_n
-
-        # Training config
-        self.epochs = args.epochs
-        self.label_smoothing = args.label_smoothing
-        # save and load model
-        self.save_folder = args.save_folder
-        self.checkpoint = args.checkpoint
-        self.continue_from = args.continue_from
-        self.model_path = args.model_path
-        # logging
-        self.print_freq = args.print_freq
-        # visualizing loss using visdom
-        self.tr_loss = torch.Tensor(self.epochs)
-        self.cv_loss = torch.Tensor(self.epochs)
-        self.visdom = args.visdom
-        self.visdom_lr = args.visdom_lr
-        self.visdom_epoch = args.visdom_epoch
-        self.visdom_id = args.visdom_id
-        if self.visdom:
-            from visdom import Visdom
-            self.vis = Visdom(env=self.visdom_id)
-            self.vis_opts = dict(title=self.visdom_id,
-                                 ylabel='Loss', xlabel='Epoch',
-                                 legend=['train loss', 'cv loss'])
-            self.vis_window = None
-            self.vis_epochs = torch.arange(1, self.epochs + 1)
-            self.optimizer.set_visdom(self.visdom_lr, self.vis)
-
-        self._reset()
-
-    def _reset(self):
-        # Reset
-        if self.continue_from:
-            print('Loading checkpoint model %s' % self.continue_from)
-            package = torch.load(self.continue_from)
-            self.model.load_state_dict(package['state_dict'])
-            self.optimizer.load_state_dict(package['optim_dict'])
-            self.start_epoch = int(package.get('epoch', 1))
-            self.tr_loss[:self.start_epoch] = package['tr_loss'][:self.start_epoch]
-            self.cv_loss[:self.start_epoch] = package['cv_loss'][:self.start_epoch]
-        else:
-            self.start_epoch = 0
-        # Create save folder
-        os.makedirs(self.save_folder, exist_ok=True)
-        self.prev_val_loss = float("inf")
-        self.best_val_loss = float("inf")
-        self.halving = False
-
-    def train(self):
-        # Train model multi-epoches
-        for epoch in range(self.start_epoch, self.epochs):
-            # Train one epoch
-            print("Training...")
-            self.model.train()  # Turn on BatchNorm & Dropout
-            start = time.time()
-            tr_avg_loss = self._run_one_epoch(epoch)
-
-            print('-' * 85)
-            print('Train Summary | End of Epoch {0} | Time {1:.2f}s | Train Loss {2:.3f}'.
-                  format(epoch + 1, time.time() - start, tr_avg_loss))
-            print('-' * 85)
-
-            # Save model each epoch
-            if epoch > 10:
-                file_path = os.path.join(
-                    self.save_folder, 'epoch-%d.pth.tar' % (epoch + 1))
-                torch.save(self.model.serialize(self.model,
-                                                self.optimizer, epoch + 1,
-                                                self.LFR_m, self.LFR_n,
-                                                tr_loss=self.tr_loss,
-                                                cv_loss=self.cv_loss),
-                           file_path)
-                print('Saving checkpoint model to %s' % file_path)
-
-            # Cross validation
-            print('Cross validation...')
-            self.model.eval()  # Turn off Batchnorm & Dropout
-            val_loss = self._run_one_epoch(epoch, cross_valid=True)
-            print('-' * 85)
-            print('Valid Summary | End of Epoch {0} | Time {1:.2f}s | '
-                  'Valid Loss {2:.3f}'.format(
-                      epoch + 1, time.time() - start, val_loss))
-            print('-' * 85)
-
-            # Save the best model
-            self.tr_loss[epoch] = tr_avg_loss
-            self.cv_loss[epoch] = val_loss
-            if val_loss < self.best_val_loss:
-                self.best_val_loss = val_loss
-                file_path = os.path.join(self.save_folder, self.model_path)
-                torch.save(self.model.serialize(self.model,
-                                                self.optimizer, epoch + 1,
-                                                self.LFR_m, self.LFR_n,
-                                                tr_loss=self.tr_loss,
-                                                cv_loss=self.cv_loss),
-                           file_path)
-                print("Find better validated model, saving to %s" % file_path)
-
-            # visualizing loss using visdom
-            if self.visdom:
-                x_axis = self.vis_epochs[0:epoch + 1]
-                y_axis = torch.stack(
-                    (self.tr_loss[0:epoch + 1], self.cv_loss[0:epoch + 1]), dim=1)
-                if self.vis_window is None:
-                    self.vis_window = self.vis.line(
-                        X=x_axis,
-                        Y=y_axis,
-                        opts=self.vis_opts,
-                    )
-                else:
-                    self.vis.line(
-                        X=x_axis.unsqueeze(0).expand(y_axis.size(
-                            1), x_axis.size(0)).transpose(0, 1),  # Visdom fix
-                        Y=y_axis,
-                        win=self.vis_window,
-                        update='replace',
-                    )
-
     def _run_one_epoch(self, epoch, cross_valid=False):
         start = time.time()
         total_loss = 0
@@ -157,9 +27,8 @@ class Solver(object):
             padded_input = padded_input.cuda()
             input_lengths = input_lengths.cuda()
             padded_target = padded_target.cuda()
-            logits = self.model(padded_input, input_lengths)
-            ctc_loss = cal_performance(logits, padded_target, smoothing=self.label_smoothing)
-            loss = ctc_loss
+            logits, len_logits = self.model(padded_input, input_lengths)
+            loss = cal_loss(logits, len_logits, padded_target, smoothing=self.label_smoothing)
             if not cross_valid:
                 self.optimizer.zero_grad()
                 loss.backward()
@@ -168,8 +37,8 @@ class Solver(object):
             total_loss += loss.item()
 
             if i % self.print_freq == 0:
-                print('Epoch {} | Iter {} | Current Loss {:.3f} | Learning Rate {:.3e} | {:.1f} ms/batch | step {}'.
-                      format(epoch + 1, i + 1, ctc_loss.item(), self.optimizer.optimizer.param_groups[0]["lr"],
+                print('Epoch {} | loss {:.3f} | lr {:.3e} | {:.1f} ms/batch | step {}'.
+                      format(epoch + 1, loss.item(), self.optimizer.optimizer.param_groups[0]["lr"],
                              1000 * (time.time() - start) / (i + 1), self.optimizer.step_num),
                       flush=True)
